+++ tests/e2e/cases/dynamic_require_type_alias/alias_a.ts
@@ -1,0 +1,2 @@
+console.log("alias a init");
+module.exports.label = "alias-a";